requirements.txt
flask
weaviate-client